Refactored note array into class and moved classes into appropriate modules
